Add meta content slot
This allows us to render PageHeadings with meta and actions:

https://tailwindui.com/components/application-ui/headings/page-headings#component-7654075f650c9f1044bfcf1b1b6ffd77

We could probably do something like we've done with actions, however for
now using content areas seem like the easier path. If we ever extract a
Meta component it might make sense to do it like actions.

We should also be aware that content slots are coming to ViewComponents,
which is yet another way of doing this.